+++ CHANGELOG.md
@@ -9,6 +9,7 @@
     - Disable [UselessParentheses](https://pmd.github.io/pmd-5.5.1/pmd-java/rules/java/unnecessary.html#UselessParentheses) because of too many false positives
     - Disable [OnlyOneReturn](https://pmd.github.io/pmd-5.5.1/pmd-java/rules/java/controversial.html#OnlyOneReturn) in favour of checkstyle ReturnCount
 * Update codenarc 0.25.1 -> 0.25.2
+* lintOptions configuration now applies to all JavaCompile tasks (not only compileJava as before) 
 
 ### 1.3.0 (2016-03-30)
 * Update checkstyle 6.14.1 -> 6.17